@@ -1,46 +1,36 @@
 class FeedsController < ApplicationController
   before_action :set_feed, only: [:update]
 
-  # GET /feeds
-  # GET /feeds.json
   def index
     @feeds = Feed.all
+
+    render json: @feeds
   end
 
-  # POST /feeds
-  # POST /feeds.json
   def create
     @feed = Feed.new(feed_params)
 
-    respond_to do |format|
-      if @feed.save
-        format.json { render :show, status: :created, location: @feed }
-      else
-        format.json { render json: @feed.errors, status: :unprocessable_entity }
-      end
+    if @feed.save
+      render json: @feed, status: :created, location: @feed
+    else
+      render json: @feed.errors, status: :unprocessable_entity
     end
   end
 
-  # PATCH/PUT /feeds/1
-  # PATCH/PUT /feeds/1.json
   def update
-    respond_to do |format|
-      if @feed.update(feed_params)
-        format.json { render :show, status: :ok, location: @feed }
-      else
-        format.json { render json: @feed.errors, status: :unprocessable_entity }
-      end
+    if @feed.update(feed_params)
+      render json: @feed
+    else
+      render json: @feed.errors, status: :unprocessable_entity
     end
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_feed
-      @feed = Feed.find(params[:id])
-    end
+  def set_feed
+    @feed = Feed.find(params[:id])
+  end
 
-    # Never trust parameters from the scary internet, only allow the white list through.
-    def feed_params
-      params.require(:feed).permit(:name, :url, :display)
-    end
+  def feed_params
+    params.require(:feed).permit(:name, :url, :display)
+  end
 end
