@@ -1,38 +1,37 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `rails
+# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170809195234) do
+ActiveRecord::Schema.define(version: 2020_07_16_192004) do
 
   create_table "entries", force: :cascade do |t|
-    t.string   "subject",                    null: false
-    t.string   "link",                       null: false
-    t.string   "data",                       null: false
-    t.boolean  "read",       default: false
-    t.date     "post_date",                  null: false
-    t.integer  "feed_id"
-    t.datetime "created_at",                 null: false
-    t.datetime "updated_at",                 null: false
-    t.boolean  "pod",        default: false
+    t.string "subject", null: false
+    t.string "link", null: false
+    t.text "data", null: false
+    t.date "post_date", null: false
+    t.boolean "pod", default: false
+    t.boolean "read", default: false
+    t.integer "feed_id", null: false
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+    t.index ["feed_id"], name: "index_entries_on_feed_id"
   end
-
-  add_index "entries", ["feed_id"], name: "index_entries_on_feed_id"
 
   create_table "feeds", force: :cascade do |t|
-    t.string   "name",                      null: false
-    t.string   "url",                       null: false
-    t.boolean  "display",    default: true
-    t.datetime "created_at",                null: false
-    t.datetime "updated_at",                null: false
+    t.string "name"
+    t.string "url"
+    t.boolean "display"
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
   end
 
+  add_foreign_key "entries", "feeds"
 end
