@@ -1,96 +1,169 @@
 require 'test_helper'
 
 class EntryTest < ActiveSupport::TestCase
-  ATOM_DATA = RSS::Parser.parse(File.read(File.join('test', 'fixtures', 'atom_feed.xml')), false)
-  RSS_DATA  = RSS::Parser.parse(File.read(File.join('test', 'fixtures', 'rss_feed.xml')), false)
-  POD_DATA  = RSS::Parser.parse(File.read(File.join('test', 'fixtures', 'rss_feed.podcast.xml')), false)
+  describe 'entry' do
+    let(:parsed_response) { RSS::Parser.parse(file, false) }
+    let(:expected_date) { Date.strptime("2017-08-08", "%Y-%m-%d") }
+    let(:data) { parsed_response.items.first }
+    let(:entry) { Entry.from_rss(data) }
 
-  def setup
-    @atom_data  = ATOM_DATA.entries.first
-    @rss_data   = RSS_DATA.items.first
-    @pod_data   = POD_DATA.items.first
-    @atom_entry = Entry.from_atom(@atom_data)
-    @rss_entry  = Entry.from_rss(@rss_data)
-    @pod_entry  = Entry.from_rss(@pod_data)
+    describe 'atom' do
+      let(:entry) { Entry.from_atom(data) }
+      let(:data) { parsed_response.entries.first }
+      let(:file) { atom_file }
 
-    @entries = {
-      atom: @atom_entry,
-      rss: @rss_entry,
-      pod: @pod_entry
-    }
+      it 'retrieves publish date' do
+        assert_equal(expected_date, entry.post_date)
+      end
 
-    super
-  end
+      it 'has a title' do
+        assert_equal("Can Real Life Compete With an Instagram Playground?", entry.subject)
+      end
 
-  test "can retrieve the publish date" do
-    expected_date = Date.strptime("2017-08-08", "%Y-%m-%d")
+      it 'has a link' do
+        assert_match(/instagram-playground-social-media/, entry.link)
+      end
 
-    @entries.values.each do |entry|
-      assert_equal(expected_date, entry.post_date)
+      it 'has no pod data' do
+        assert_nil(entry.data)
+        assert_equal(false, entry.pod?)
+      end
+    end
+
+    describe 'rss' do
+      let(:file) { rss_file }
+
+      it 'retrieves publish date' do
+        assert_equal(expected_date, entry.post_date)
+      end
+
+      it 'has a title' do
+        assert_equal("Trump Picks A Favorite In Alabama’s GOP Senate Primary", entry.subject)
+      end
+
+      it 'has a link' do
+        assert_match(/trump-picks-a-favorite-in-alabamas-gop-senate-primary/, entry.link)
+      end
+
+      it 'has no pod data' do
+        assert_nil(entry.data)
+        assert_equal(false, entry.pod?)
+      end
+    end
+
+    describe 'rss pod' do
+      let(:file) { rss_podcast_file }
+
+      it 'retrieves publish date' do
+        assert_equal(expected_date, entry.post_date)
+      end
+
+      it 'has a title' do
+        assert_equal("The 25 Greatest Patriots Wins of the Brady-Belichick Era (Ep. 245)", entry.subject)
+      end
+
+      it 'has a link' do
+        assert_match(/71e589d6-92e5-4536-83ca-af3cbbd1bfc9/, entry.link)
+      end
+
+      it 'can retrieve the length' do
+        assert_match(/01:52:37/, entry.data[:length])
+        assert_equal(true, entry.pod?)
+      end
     end
   end
 
-  test "can retrieve the title" do
-    assert_equal("Can Real Life Compete With an Instagram Playground?", @entries[:atom].subject)
-    assert_equal("The 25 Greatest Patriots Wins of the Brady-Belichick Era (Ep. 245)", @entries[:pod].subject)
-    assert_equal("Trump Picks A Favorite In Alabama’s GOP Senate Primary", @entries[:rss].subject)
+  describe '#subject' do
+    let(:entry) { Entry.new.set_subject(title, description) }
+
+    describe 'short title' do
+      let(:title) { 'Hi' }
+
+      describe 'with description' do
+        let(:description) { 'descripty description' }
+
+        it 'adds description' do
+          assert_equal([title, description].join(': '), entry.subject)
+        end
+      end
+
+      describe 'with long description' do
+        let(:description) { SecureRandom.hex(Entry::MAX_COMPOUND_SUBJECT_SIZE + (Entry::MAX_COMPOUND_SUBJECT_SIZE * 2)) }
+
+        it 'adds description' do
+          assert_equal([title, description].join(': ')[0..Entry::MAX_COMPOUND_SUBJECT_SIZE + 1], entry.subject)
+        end
+      end
+
+      describe 'no description' do
+        let(:description) { nil }
+
+        it 'adds description' do
+          assert_equal(title, entry.subject)
+        end
+      end
+    end
+
+    describe 'no title' do
+      let(:title) { nil }
+
+      describe 'with description' do
+        let(:description) { 'descripty description' }
+
+        it 'adds description' do
+          assert_equal(description, entry.subject)
+        end
+      end
+
+      describe 'no description' do
+        let(:description) { nil }
+
+        it 'adds description' do
+          assert_equal('nil Title and nil Description', entry.subject)
+        end
+      end
+    end
   end
 
-  test "can add description if title is short" do
-    rss_data   = RSS_DATA.items.last
-    rss_entry  = Entry.from_rss(rss_data)
-    assert_equal("Democrats Divided: As the Democratic Party increasingly embraces single payer healthcare and Clinton weighs in on her primary", rss_entry.subject)
-  end
+  describe '#exists?' do
+    let(:entry) { entries(:one) }
 
-  test "can handle short title and long word description" do
-    long_description = SecureRandom.hex(Entry::MAX_COMPOUND_SUBJECT_SIZE + (Entry::MAX_COMPOUND_SUBJECT_SIZE * 2))
-    rss_entry  = Entry.new.set_subject("Hi", long_description)
-    expected   = "Hi: #{long_description[0..Entry::MAX_COMPOUND_SUBJECT_SIZE - "Hi:".size]}"
-    assert_equal(expected, rss_entry.subject)
-  end
+    before do
+      entry.update!(link: "https://blah.com/abc123/my-article/")
+    end
 
-  test "can handle no description and short title" do
-    rss_entry  = Entry.new.set_subject("Hi", nil)
-    assert_equal("Hi", rss_entry.subject)
-  end
+    describe 'already exists' do
+      it 'matches exactly' do
+        assert_equal(true, Entry.new({link: "https://blah.com/abc123/my-article/"}).exists?)
+      end
 
-  test "can handle no title and short description" do
-    rss_entry  = Entry.new.set_subject(nil, "Hi")
-    assert_equal("Hi", rss_entry.subject)
-  end
+      it 'only differs in http(s)' do
+        assert_equal(true, Entry.new({link: "http://blah.com/abc123/my-article/"}).exists?)
+      end
 
-  test "can handle no title and no description" do
-    rss_entry  = Entry.new.set_subject(nil, nil)
-    assert_equal("nil Title and nil Description", rss_entry.subject)
-  end
+      it 'only differs in www' do
+        assert_equal(true, Entry.new({link: "https://www.blah.com/abc123/my-article/"}).exists?)
+      end
 
-  test "can retrieve the link" do
-    assert_match(/instagram-playground-social-media/, @entries[:atom].link)
-    assert_match(/71e589d6-92e5-4536-83ca-af3cbbd1bfc9/, @entries[:pod].link)
-    assert_match(/trump-picks-a-favorite-in-alabamas-gop-senate-primary/, @entries[:rss].link)
-  end
+      it 'has different hostname' do
+        assert_equal(true, Entry.new({link: "https://blerg.com/abc123/my-article/"}).exists?)
+      end
 
-  test "can retrieve the length of podcasts" do
-    assert_nil(@entries[:atom].data)
-    assert_equal(false, @entries[:atom].pod)
-    assert_match(/01:52:37/, @entries[:pod].data[:length])
-    assert(@entries[:pod].pod)
-    assert_nil(@entries[:rss].data)
-    assert_equal(false, @entries[:rss].pod)
-  end
+      it 'has extra slashes' do
+        assert_equal(true, Entry.new({link: "https://blerg.com//abc123/my-article/"}).exists?)
+      end
 
-  test "can tell if already exists" do
-    data = {subject: "hello", link: "https://blah.com/abc123/my-article/", post_date: entries(:one).post_date}
+      it 'has no slash on the end' do
+        assert_equal(true, Entry.new({link: "https://blerg.com/abc123/my-article"}).exists?)
+      end
 
-    Entry.new(data).save!
+      it 'has no hostname' do
+        assert_equal(true, Entry.new({link: "abc123/my-article"}).exists?)
+      end
+    end
 
-    assert_equal(true, Entry.new({link: "https://blah.com/abc123/my-article/"}).exists?)
-    assert_equal(true, Entry.new({link: "http://blah.com/abc123/my-article/"}).exists?)
-    assert_equal(true, Entry.new({link: "https://www.blah.com/abc123/my-article/"}).exists?)
-    assert_equal(true, Entry.new({link: "https://blerg.com/abc123/my-article/"}).exists?)
-    assert_equal(true, Entry.new({link: "https://blerg.com//abc123/my-article/"}).exists?)
-    assert_equal(true, Entry.new({link: "https://blerg.com/abc123/my-article"}).exists?)
-    assert_equal(true, Entry.new({link: "abc123/my-article"}).exists?)
-    assert_equal(false, Entry.new({link: "https://blah.com/abc123/my-article-hi/"}).exists?)
+    it 'is different' do
+      assert_equal(false, Entry.new({link: "https://blah.com/abc123/my-article-hi/"}).exists?)
+    end
   end
 end
